Add quantification of the influence of reef features on cluster assigns
Added quantification of cluster-level ~ reef factor affects (such as depth and connectivity) via a multinomial logistic regression (provided by Econometrics.jl pkg). This is an extension of binomial logistic regression.

This regression is performed individually for each GCM. Only ACCESS-ESM1-5 results are displayed in the paper.

Added table of model coefficients to the results section of paper but have not added these values into text or discussed values yet.