Streamlining time series clustering description
Removing some unnecessary detail

Implementation details, e.g., about size and shape of matrices, are not necessary for this kind of paper. If needed, they can be explained in supplementaries.

Add section in supplementary for expanded info on time series clustering (may remove, as it's described in cited paper...).

Also some comments for @eco-ben